refactor(single-service): split code, reuse component and fix timeline cta button
