+++ src/common/components/waterButton.tsx
@@ -1,0 +1,8 @@
+import { Button, theme } from '@appquality/unguess-design-system';
+import styled from 'styled-components';
+
+export const WaterButton = styled(Button)``;
+
+WaterButton.defaultProps = {
+  themeColor: theme.palette.water[600],
+};
