--- src/common/theme-variant.ts
@@ -1,11 +1,0 @@
-import { theme } from '@appquality/unguess-design-system';
-
-const customTheme = {
-  ...theme,
-  colors: {
-    ...theme.colors,
-    primaryHue: theme.palette.water[600],
-  },
-};
-
-export default customTheme;
