+++ src/pages/ExpressWizard/steps/forms/WhatForm.tsx
@@ -2,7 +2,7 @@ import { FormikProps } from 'formik';
 import { useTranslation } from 'react-i18next';
 import { WizardModel } from '../../wizardModel';
 import { WhatStep } from '../what';
-import { WaterButton } from '../../waterButton';
+import { WaterButton } from '../../../../common/components/waterButton';
 import { WizardButtonsProps } from './types';
 
 export const WhatForm = (props: FormikProps<WizardModel>) => (
