--- src/pages/ExpressWizard/waterButton.tsx
@@ -1,9 +1,0 @@
-import { Button } from '@appquality/unguess-design-system';
-import styled from 'styled-components';
-import customTheme from 'src/common/theme-variant';
-
-export const WaterButton = styled(Button)``;
-
-WaterButton.defaultProps = {
-  theme: customTheme,
-};
