+++ src/pages/Service/ServiceCta.tsx
@@ -1,0 +1,6 @@
+import { WaterButton } from 'src/common/components/waterButton';
+import styled from 'styled-components';
+
+export const ServiceCta = styled(WaterButton)`
+  margin: ${({ theme }) => theme.space.md} 0;
+`;
